deabstract print_next_state
